Implement features

+++ src/main/java/com/accenture/codingtest/springbootcodingtest/service/impl/UserServiceImpl.java
@@ -1,17 +1,19 @@
 package com.accenture.codingtest.springbootcodingtest.service.impl;
 
-import com.accenture.codingtest.springbootcodingtest.ErrorDsc;
 import com.accenture.codingtest.springbootcodingtest.dto.UserDto;
 import com.accenture.codingtest.springbootcodingtest.entity.Project;
 import com.accenture.codingtest.springbootcodingtest.entity.Task;
 import com.accenture.codingtest.springbootcodingtest.entity.User;
+import com.accenture.codingtest.springbootcodingtest.enums.Role;
 import com.accenture.codingtest.springbootcodingtest.exception.ProjectNotFound;
 import com.accenture.codingtest.springbootcodingtest.exception.TaskNotFoundException;
+import com.accenture.codingtest.springbootcodingtest.exception.UnAuthorized;
 import com.accenture.codingtest.springbootcodingtest.exception.UserNotFoundException;
 import com.accenture.codingtest.springbootcodingtest.repository.ProjectRepository;
 import com.accenture.codingtest.springbootcodingtest.repository.TaskRepository;
 import com.accenture.codingtest.springbootcodingtest.repository.UserRepository;
 import com.accenture.codingtest.springbootcodingtest.service.UserService;
+import com.accenture.codingtest.springbootcodingtest.util.ErrorDsc;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.util.ReflectionUtils;
@@ -41,20 +43,27 @@ public class UserServiceImpl implements UserService {
      * @throws ProjectNotFound
      */
     @Override
-    public String createUser(UserDto userDto) throws TaskNotFoundException, ProjectNotFound {
-        User user = new User();
-        user.setUserName(userDto.getUserName());
-        user.setPassword(userDto.getPassword());
-        Project project = projectRepository.findById(userDto.getProjectId()).orElseThrow(() -> new ProjectNotFound(ErrorDsc.ERR_DSC_PROJECT_NOT_FOUND));
+    public String createUser(UserDto userDto, String roleId) throws TaskNotFoundException, ProjectNotFound, UnAuthorized {
 
-        List<User> users = new ArrayList<>();
-        users.add(user);
-        Task task = taskRepository.findById(userDto.getTaskId()).orElseThrow(() -> new TaskNotFoundException(ErrorDsc.ERR_DSC_TASK_NOT_FOUND));
-        user.setTask(task);
-        project.setUsers(users);
+        if (roleId.equalsIgnoreCase(Role.ADMIN.toString())) {
+            User user = new User();
+            user.setUserName(userDto.getUserName());
+            user.setPassword(userDto.getPassword());
+            Project project = projectRepository.findById(userDto.getProjectId()).orElseThrow(() -> new ProjectNotFound(ErrorDsc.ERR_DSC_PROJECT_NOT_FOUND));
 
-        userRepository.save(user);
-        return "User created successfully";
+            List<User> users = new ArrayList<>();
+            users.add(user);
+            Task task = taskRepository.findById(userDto.getTaskId()).orElseThrow(() -> new TaskNotFoundException(ErrorDsc.ERR_DSC_TASK_NOT_FOUND));
+            user.setTask(task);
+            project.setUsers(users);
+
+            userRepository.save(user);
+            return "User created successfully";
+
+        } else {
+            throw new UnAuthorized(ErrorDsc.ERR_DSC_UNAUTHORIZED);
+        }
+
     }
 
     /**
@@ -63,8 +72,13 @@ public class UserServiceImpl implements UserService {
      * @return
      */
     @Override
-    public List<User> getUsers() {
-        return userRepository.findAll();
+    public List<User> getUsers(String roleId) throws UnAuthorized {
+        if (roleId.equalsIgnoreCase(Role.ADMIN.toString())) {
+            return userRepository.findAll();
+        } else {
+            throw new UnAuthorized(ErrorDsc.ERR_DSC_UNAUTHORIZED);
+        }
+
     }
 
     /**
@@ -75,8 +89,13 @@ public class UserServiceImpl implements UserService {
      * @throws UserNotFoundException
      */
     @Override
-    public User getUser(UUID userId) throws UserNotFoundException {
-        return userRepository.findById(userId).orElseThrow(() -> new UserNotFoundException(ErrorDsc.ERR_DSC_USER_NOT_FOUND));
+    public User getUser(UUID userId, String roleId) throws UserNotFoundException, UnAuthorized {
+        if (roleId.equalsIgnoreCase(Role.ADMIN.toString())) {
+            return userRepository.findById(userId).orElseThrow(() -> new UserNotFoundException(ErrorDsc.ERR_DSC_USER_NOT_FOUND));
+        } else {
+            throw new UnAuthorized(ErrorDsc.ERR_DSC_UNAUTHORIZED);
+        }
+
     }
 
     /**
@@ -88,11 +107,15 @@ public class UserServiceImpl implements UserService {
      * @throws UserNotFoundException
      */
     @Override
-    public User updateUser(UserDto userDto, UUID userId) throws UserNotFoundException {
-        User user = userRepository.findById(userId).orElseThrow(() -> new UserNotFoundException(ErrorDsc.ERR_DSC_USER_NOT_FOUND));
-        user.setUserName(userDto.getUserName());
-        user.setPassword(userDto.getPassword());
-        return userRepository.save(user);
+    public User updateUser(UserDto userDto, UUID userId, String roleId) throws UserNotFoundException, UnAuthorized {
+        if (roleId.equalsIgnoreCase(Role.ADMIN.toString())) {
+            User user = userRepository.findById(userId).orElseThrow(() -> new UserNotFoundException(ErrorDsc.ERR_DSC_USER_NOT_FOUND));
+            user.setUserName(userDto.getUserName());
+            user.setPassword(userDto.getPassword());
+            return userRepository.save(user);
+        } else {
+            throw new UnAuthorized(ErrorDsc.ERR_DSC_UNAUTHORIZED);
+        }
     }
 
     /**
@@ -104,16 +127,20 @@ public class UserServiceImpl implements UserService {
      * @throws UserNotFoundException
      */
     @Override
-    public User updateUserPartially(Map<String, Object> fields, UUID userId) throws UserNotFoundException {
-        User user = userRepository.findById(userId).orElseThrow(() -> new UserNotFoundException(ErrorDsc.ERR_DSC_USER_NOT_FOUND));
+    public User updateUserPartially(Map<String, Object> fields, UUID userId, String roleId) throws UserNotFoundException, UnAuthorized {
+        if (roleId.equalsIgnoreCase(Role.ADMIN.toString())) {
+            User user = userRepository.findById(userId).orElseThrow(() -> new UserNotFoundException(ErrorDsc.ERR_DSC_USER_NOT_FOUND));
 
-        fields.forEach((requestField, value) -> {
-            Field field = ReflectionUtils.findField(User.class, requestField);
-            field.setAccessible(true);
-            ReflectionUtils.setField(field, user, value);
-        });
+            fields.forEach((requestField, value) -> {
+                Field field = ReflectionUtils.findField(User.class, requestField);
+                field.setAccessible(true);
+                ReflectionUtils.setField(field, user, value);
+            });
 
-        return userRepository.save(user);
+            return userRepository.save(user);
+        } else {
+            throw new UnAuthorized(ErrorDsc.ERR_DSC_UNAUTHORIZED);
+        }
     }
 
     /**
@@ -124,9 +151,13 @@ public class UserServiceImpl implements UserService {
      * @throws UserNotFoundException
      */
     @Override
-    public String deleteUser(UUID userId) throws UserNotFoundException {
-        User user = userRepository.findById(userId).orElseThrow(() -> new UserNotFoundException(ErrorDsc.ERR_DSC_USER_NOT_FOUND));
-        userRepository.deleteById(user.getId());
-        return "User successfully deleted";
+    public String deleteUser(UUID userId, String roleId) throws UserNotFoundException, UnAuthorized {
+        if (roleId.equalsIgnoreCase(Role.ADMIN.toString())) {
+            User user = userRepository.findById(userId).orElseThrow(() -> new UserNotFoundException(ErrorDsc.ERR_DSC_USER_NOT_FOUND));
+            userRepository.deleteById(user.getId());
+            return "User successfully deleted";
+        } else {
+            throw new UnAuthorized(ErrorDsc.ERR_DSC_UNAUTHORIZED);
+        }
     }
 }
